refactor history methods
